Add solutions if passed
